Setting up animatoin on signup vs login form. Far from perfect, but getting better
