+++ src/scss/components/_forms.scss
@@ -1,12 +1,16 @@
 @use "../utilities/variables" as variables;
 
+.moving-main {
+  transition: 500ms ease-in-out;
+}
+
 .login-section {
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
   position: relative;
-  top: 35vh;
+  top: 25vh;
 
   form {
     display: flex;
@@ -29,10 +33,45 @@
       }
     }
   }
+
+  .show-signup-form-button {
+    margin-top: 7vh;
+  }
 }
 
 .invalid-feedback {
   font-size: 0.8rem;
   margin: 0;
   color: map-get($map: variables.$colors, $key: jetred);
+}
+
+.signup-secton {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+  top: 100vh;
+
+  form {
+    display: flex;
+    flex-direction: column;
+    width: fit-content;
+
+    fieldset {
+      border: none;
+      display: flex;
+      flex-direction: column;
+
+      div {
+        height: 100%;
+        display: flex;
+        flex-direction: column;
+
+        label {
+          margin-top: 15px;
+        }
+      }
+    }
+  }
 }
